Final_update
Todos los arreglos hechos y ya pasa Francinette en normal y estricto.

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: alcarden <alcarden@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/20 12:06:57 by alcarden          #+#    #+#             */
-/*   Updated: 2023/04/20 12:14:57 by alcarden         ###   ########.fr       */
+/*   Updated: 2023/04/27 18:40:32 by alcarden         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*str;
 
 	i = 0;
